bugfix: change transporting Header to jwtTokenprovider to transporting pure accesstoken to jwtTokenprovider.
I'm sorry to write in English. cuz my laptop is enable to use Korean.

In this issue, I just solve the problem that getUserIdFromAccessToken recepts 'Bearer <Token>'.
In other words, UserIdResolver passes the Bearer prefix from the Authorization header to JwtTokenProvider.getUserIdFromAccessToken() without removing it.
So, it just causes JWT strings may not contain whitespace.

In this situatioin, I just solve this bug with adding extractToken method in UserIdResolver, So I just let remove string Bearer in getUserIdFromAccessToken method parameter
But in this endpoint, it calls getUserIdFromAccessToken 2 times. First in jwtInterceptor, and in UserIdResolver.
It can't be proper clean code. So If you want any refactoring then do it, or give me another feedback.